+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next release (tbd)
 
  - fix: add timeout of 2.5s when polling and log error as JSON
+ - fix: use `.on('get', fn)` if `.onGet()` is not available (to support Homebridge v1.2.5)
 
 ## v1.1.0 (2021-06-23)